requirements.txt
altgraph==0.17.5
boto3==1.42.30
botocore==1.42.30
humanize==4.15.0
jmespath==1.0.1
packaging==25.0
pyaes==1.6.1
pyasn1==0.6.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pymediainfo-pyrofork==6.0.2
pyrofork==2.3.69
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rsa==4.9.1
s3transfer==0.16.0
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
telegram-download-chat==0.5.4
Telethon==1.42.0
TgCrypto==1.2.5
urllib3==2.6.3
